2.4.10A

+++ actions.php
@@ -2,9 +2,9 @@
 
 include_once __DIR__.'/core.php';
 
-use Models\Checklist\Check;
 use Models\Note;
-use Models\User;
+use Modules\Checklists\Check;
+use Modules\Checklists\Checklist;
 
 if (empty($structure) || empty($structure['enabled'])) {
     die(tr('Accesso negato'));
@@ -78,6 +78,18 @@ elseif (filter('op') == 'download_file') {
     download($upload_dir.'/'.$rs[0]['filename'], $rs[0]['original']);
 }
 
+// Modifica nome della categoria degli allegati
+elseif (filter('op') == 'upload_category') {
+    $category = post('category');
+    $name = post('name');
+
+    $uploads = $structure->uploads($id_record)->where('category', $category);
+    foreach ($uploads as $upload) {
+        $upload->category = $name;
+        $upload->save();
+    }
+}
+
 // Validazione dati
 elseif (filter('op') == 'validate') {
     // Lettura informazioni di base
@@ -107,6 +119,17 @@ elseif (filter('op') == 'add_nota') {
     flash()->info(tr('Nota interna aggiunta correttamente!'));
 }
 
+// Rimozione data di notifica dalla nota interna
+elseif (filter('op') == 'notification_nota') {
+    $id_nota = post('id_nota');
+    $nota = Note::find($id_nota);
+
+    $nota->notification_date = null;
+    $nota->save();
+
+    flash()->info(tr('Data di notifica rimossa dalla nota interna!'));
+}
+
 // Rimozione nota interna
 elseif (filter('op') == 'delete_nota') {
     $id_nota = post('id_nota');
@@ -117,22 +140,44 @@ elseif (filter('op') == 'delete_nota') {
     flash()->info(tr('Nota interna aggiunta correttamente!'));
 }
 
-// Rimozione checklist
+// Clonazione di una checklist
+elseif (filter('op') == 'clone_checklist') {
+    $content = post('content');
+    $checklist_id = post('checklist');
+
+    $users = post('assigned_users');
+    $users = array_clean($users);
+
+    $group_id = post('group_id');
+
+    $checklist = Checklist::find($checklist_id);
+    $checklist->copia($user, $id_record, $users, $group_id);
+}
+
+// Aggiunta check alla checklist
 elseif (filter('op') == 'add_check') {
     $content = post('content');
     $parent_id = post('parent') ?: null;
 
-    $assigned_user = User::find(post('assigned_user'));
+    $users = post('assigned_users');
+    $users = array_clean($users);
 
-    $check = Check::build($user, $assigned_user, $structure, $id_record, $content, $parent_id);
+    $group_id = post('group_id');
+
+    $check = Check::build($user, $structure, $id_record, $content, $parent_id);
+    $check->setAccess($users, $group_id);
 }
 
-// Rimozione checklist
+// Rimozione di un check della checklist
 elseif (filter('op') == 'delete_check') {
     $check_id = post('check_id');
     $check = Check::find($check_id);
 
-    $check->delete();
+    if (!empty($check) && $check->user->id == $user->id) {
+        $check->delete();
+    } else {
+        flash()->error(tr('Impossibile eliminare il check!'));
+    }
 }
 
 // Gestione check per le checklist
@@ -140,13 +185,21 @@ elseif (filter('op') == 'toggle_check') {
     $check_id = post('check_id');
     $check = Check::find($check_id);
 
-    if (!empty($check)) {
-        $check->checked_at = $check->checked_at ? null : date('Y-m-d H:i:s');
-        $check->save();
+    if (!empty($check) && $check->assignedUsers->pluck('id')->search($user->id) !== false) {
+        $check->toggleCheck($user);
+    } else {
+        flash()->error(tr('Impossibile cambiare lo stato del check!'));
+    }
+}
 
-        echo json_encode([
-            'checked_at' => timestampFormat($check->checked_at) ?: null,
-        ]);
+// Gestione ordine per le checklist
+elseif (filter('op') == 'sort_checks') {
+    $ids = explode(',', $_POST['order']);
+    $order = 0;
+
+    foreach ($ids as $id) {
+        $dbo->query('UPDATE `zz_checks` SET `order` = '.prepare($order).' WHERE id = '.prepare($id));
+        ++$order;
     }
 }
